Editor: Move Editor API inside the plugin
