--- spyder/api/editor.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# Copyright © Spyder Project Contributors
-# Licensed under the terms of the MIT License
-# (see spyder/__init__.py for details)
-
-"""
-This module contains the code editor API.
-"""
-
-from spyder.plugins.editor.api.decoration import TextDecoration
