+++ spyder/plugins/editor/api/manager.py
