+++ spyder/plugins/editor/api/panel.py
@@ -26,7 +26,7 @@ from qtpy.QtGui import QBrush, QColor, QPen, QPainter
 from qtpy.QtCore import Qt, QRect
 
 # Local imports
-from spyder.api.editorextension import EditorExtension
+from spyder.plugins.editor.api.editorextension import EditorExtension
 from spyder.plugins.outlineexplorer.api import is_cell_header
 
 logger = logging.getLogger(__name__)
