+++ spyder/plugins/editor/extensions/closequotes.py
@@ -12,7 +12,7 @@ import re
 from qtpy.QtGui import QTextCursor
 
 # Local imports
-from spyder.api.editorextension import EditorExtension
+from spyder.plugins.editor.api.editorextension import EditorExtension
 
 
 def unmatched_quotes_in_line(text):
