+++ spyder/plugins/editor/extensions/snippets.py
@@ -22,8 +22,8 @@ except Exception:
     rtree_available = False
 
 # Local imports
+from spyder.plugins.editor.api.editorextension import EditorExtension
 from spyder.py3compat import to_text_string
-from spyder.api.editorextension import EditorExtension
 from spyder.utils.snippets.ast import build_snippet_ast, nodes, tokenize
