+++ spyder/plugins/editor/widgets/codeeditor/codeeditor.py
@@ -42,9 +42,9 @@ from spyder_kernels.utils.dochelpers import getobj
 
 
 # Local imports
-from spyder.api.panel import Panel
 from spyder.config.base import _, running_under_pytest
 from spyder.plugins.editor.api.decoration import TextDecoration
+from spyder.plugins.editor.api.panel import Panel
 from spyder.plugins.editor.extensions import (CloseBracketsExtension,
                                               CloseQuotesExtension,
                                               DocstringWriterExtension,
